Similarity panel and configurator
Similarity Panel and similarity configurator classes has been migrated to typescript. Due to the migration, we created a  nano-markdown  module definition for typescript and added the bootstrap type module to the package.json
Additionally, we created a templateManager class that hides handlebars from the rest of the modules. For this reason , table comparator has been modified
We removed unnecesary files
