+++ src/js/types/nano-markdown.d.ts
@@ -1,0 +1,1 @@
+declare module 'nano-markdown';
